# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentPay API
  version: 0.4.0
  description: |
    Seller commerce gateway, approval, transaction, and dispute API. AgentPay is a working name.
    Version 0.4.0 is the locked production target contract for M7.1. The current
    M7 Go runtime remains on the documented development compatibility baseline
    until the dependent LCH implementation tasks land. Production deployment
    must not advertise or partially enable this target contract before every
    authorization and persistence precondition is implemented.
    Seller project keys are proprietary bootstrap credentials accepted only by
    POST /v1/integration-access-tokens; this is not an OAuth token endpoint.
    Lean V1 project-key installations use the required AgentPay connector.
    Direct remote MCP OAuth clients are deferred. Ordinary MCP requests use
    short-lived ES256 access tokens. Browser
    purchase sessions and buyer-agent keys are distinct from seller credentials.
    Production MCP commercial mutations additionally require a one-time grant
    issued through the authenticated seller browser/BFF boundary; caller-
    asserted confirmation fields are development compatibility only.
  x-agentpay-contract-stage: target-state
  x-agentpay-runtime-baseline: m7-development-preview
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://{apiHost}
    description: Environment-specific API Gateway origin
    variables:
      apiHost:
...
paths:
...
  /v1/intents:
    post:
      summary: Create an immutable purchase intent
      tags: [Intents]
      operationId: createPurchaseIntent
      security: [{ buyerAgentKey: [] }, { browserPurchaseCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/PurchaseCsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateIntentRequest' }
      responses:
        '201':
          description: Immutable purchase intent created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseIntent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410': { $ref: '#/components/responses/SellerInactive' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /v1/intents/{intentId}:
    get:
      summary: Get a purchase intent
      tags: [Intents]
      operationId: getPurchaseIntent
      security: [{ buyerAgentKey: [] }, { browserPurchaseCookie: [] }, { sellerBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/IntentId' }]
      responses:
        '200':
          description: Purchase intent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseIntent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
...
  /store/{slug}/manifest.json:
    get:
      summary: Get a storefront manifest
      tags: [Storefront]
      operationId: getStorefrontManifest
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, pattern: '^[a-z0-9-]+$' }
      responses:
        '200':
          description: Short-lived signed active storefront discovery document
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedStorefrontManifest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410':
          description: Known storefront is inactive; the signed tombstone contains no products
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedStorefrontTombstone' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /store/{slug}/llms.txt:
    get:
      summary: Get the storefront agent-discovery document
      tags: [Storefront]
      operationId: getStorefrontLlmsText
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, pattern: '^[a-z0-9-]+$' }
      responses:
        '200':
          description: Plain-text agent discovery document
          content:
            text/plain:
              schema: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { $ref: '#/components/responses/SellerInactive' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /v1/storefronts/{sellerSlug}/products/{productSlug}:
    get:
      summary: Resolve one active public product by seller and product slug
      tags: [Storefront]
      operationId: getPublicProduct
      security: []
      parameters:
        - { $ref: '#/components/parameters/SellerSlug' }
        - { $ref: '#/components/parameters/ProductSlug' }
      responses:
        '200':
          description: Signed public product document; discovery does not authorize a purchase
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedPublicProductDocument' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { $ref: '#/components/responses/SellerInactive' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
...
  /pay/{slug}/{proxyPath}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
      - name: proxyPath
        in: path
        required: true
        description: Greedy path parameter in API Gateway deployment
        schema: { type: string }
      - name: X-AgentPay-Intent-Id
        in: header
        required: true
        schema: { type: string }
      - name: X-AgentPay-Approval
        in: header
        required: false
        schema: { type: string }
      - name: PAYMENT-SIGNATURE
        in: header
        required: false
        schema: { type: string }
      - { $ref: '#/components/parameters/PurchaseCsrfToken' }
    get:
      summary: Purchase a configured GET resource
      tags: [Proxy]
      operationId: getPaidResource
      security: [{ buyerAgentKey: [] }, { browserPurchaseCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/ProxiedResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410': { $ref: '#/components/responses/SellerInactive' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '428': { $ref: '#/components/responses/ApprovalRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
    post:
      summary: Purchase a configured POST resource
      tags: [Proxy]
      operationId: postPaidResource
      security: [{ buyerAgentKey: [] }, { browserPurchaseCookie: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema: {}
      responses:
        '200': { $ref: '#/components/responses/ProxiedResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410': { $ref: '#/components/responses/SellerInactive' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '428': { $ref: '#/components/responses/ApprovalRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
components:
  securitySchemes:
    sellerBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Seller-scoped service credential derived from the authenticated web session by the trusted web/BFF boundary; it is never stored in browser JavaScript or accepted as a substitute for ownership checks.
    projectKey:
      type: apiKey
      in: header
      name: X-AgentPay-Project-Key
      description: Proprietary seller project bootstrap credential accepted only by the integration access-token exchange.
    buyerAgentKey:
      type: apiKey
      in: header
      name: X-AgentPay-Agent-Key
      description: Buyer-agent credential; never a seller project key.
    browserPurchaseCookie:
      type: apiKey
      in: cookie
      name: __Host-agentpay_purchase
      description: Opaque server-side browser grant. Commerce authority expires separately from longer-lived read and remediation access.
    approvalCookie:
      type: apiKey
      in: cookie
      name: __Host-agentpay_approvals
      description: Opaque browser grant set created by exchanging one or more approval invitation fragment tokens.
    stripeWebhookSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: Stripe-generated signature verified against the exact raw request body and the environment-specific webhook endpoint secret.
  parameters:
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' } }
...
    IntentId: { name: intentId, in: path, required: true, schema: { type: string, pattern: '^int_[A-Za-z0-9]+$' } }
...
    SellerSlug: { name: sellerSlug, in: path, required: true, schema: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', minLength: 3, maxLength: 48 } }
    ProductSlug: { name: productSlug, in: path, required: true, schema: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', minLength: 3, maxLength: 80 } }
...
    PurchaseCsrfToken:
      name: X-AgentPay-CSRF
      in: header
      required: false
      description: Required for every state-changing request, including paid GET execution, when browserPurchaseCookie is used; ignored for buyer-agent key authentication. It must match the __Host-agentpay_purchase_csrf double-submit cookie and the request Origin must be allowlisted.
      schema: { type: string, minLength: 32, maxLength: 256 }
...
  schemas:
...
    CreateIntentRequest:
      type: object
      additionalProperties: false
      required: [routeId, requestBodyHash, maximumAmount]
      properties:
        routeId: { type: string }
        requestBodyHash: { $ref: '#/components/schemas/Sha256' }
        maximumAmount: { $ref: '#/components/schemas/AtomicAmount' }
    PurchaseIntent:
      type: object
      additionalProperties: false
      required: [intentId, routeId, sellerId, buyerId, purchaseChannel, productDisplayName, productSlug, paymentDestinationId, payTo, requestMethod, requestPath, amount, asset, network, requestBodyHash, intentHash, maximumAmount, requiresApproval, status, expiresAt, createdAt]
      properties:
        intentId: { type: string }
        routeId: { type: string }
        sellerId: { type: string }
        buyerId: { type: string, description: Buyer-agent subject for agent purchases; browser:<purchaseSessionId> for browser purchases. }
        purchaseSessionId: { type: [string, 'null'], pattern: '^bps_[A-Za-z0-9]+$' }
        purchaseChannel: { type: string, enum: [agent, browser] }
        productDisplayName: { type: string, minLength: 1, maxLength: 120 }
        productSlug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        paymentDestinationId: { type: string, pattern: '^dst_[A-Za-z0-9]+$' }
        payTo: { type: string, minLength: 1, maxLength: 256 }
        requestMethod: { type: string, enum: [GET, POST] }
        requestPath: { type: string, pattern: '^/[A-Za-z0-9/_-]+$' }
        amount: { $ref: '#/components/schemas/AtomicAmount' }
        asset: { type: string }
        network: { type: string }
        requestBodyHash: { $ref: '#/components/schemas/Sha256' }
        intentHash: { $ref: '#/components/schemas/Sha256' }
        maximumAmount: { $ref: '#/components/schemas/AtomicAmount' }
        requiresApproval: { type: boolean }
        status: { type: string, enum: [ready, approval_pending, approved, expired, executed] }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
      allOf:
        - if:
            properties: { purchaseChannel: { const: browser } }
            required: [purchaseChannel]
          then:
            required: [purchaseSessionId]
            properties:
              purchaseSessionId: { type: string, pattern: '^bps_[A-Za-z0-9]+$' }
              buyerId: { type: string, pattern: '^browser:bps_[A-Za-z0-9]+$' }
        - if:
            properties: { purchaseChannel: { const: agent } }
            required: [purchaseChannel]
          then:
            properties:
              purchaseSessionId: { type: 'null' }
...
    StorefrontManifest:
      type: object
      additionalProperties: false
      required: [schemaVersion, sellerId, seller, availability, publicationRevision, issuedAt, expiresAt, canonicalOrigin, products]
      properties:
        schemaVersion: { type: string, const: agentpay.discovery.v1 }
        sellerId: { type: string, pattern: '^sel_[A-Za-z0-9]+$' }
        seller:
          type: object
          additionalProperties: false
          required: [name, slug]
          properties:
            name: { type: string }
            slug: { type: string }
        availability: { type: string, const: active }
        publicationRevision: { type: integer, minimum: 1 }
        issuedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        canonicalOrigin: { type: string, format: uri }
        products:
          type: array
          items: { $ref: '#/components/schemas/PublicProduct' }
...
    PublicProductDocument:
      type: object
      additionalProperties: false
      required: [schemaVersion, sellerId, sellerSlug, publicationRevision, issuedAt, expiresAt, canonicalOrigin, product]
      properties:
        schemaVersion: { type: string, const: agentpay.discovery.v1 }
        sellerId: { type: string, pattern: '^sel_[A-Za-z0-9]+$' }
        sellerSlug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        publicationRevision: { type: integer, minimum: 1 }
        issuedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        canonicalOrigin: { type: string, format: uri }
        product: { $ref: '#/components/schemas/PublicProduct' }
    StorefrontTombstone:
      type: object
      additionalProperties: false
      required: [schemaVersion, sellerId, sellerSlug, availability, reason, publicationRevision, issuedAt, expiresAt, canonicalOrigin]
      properties:
        schemaVersion: { type: string, const: agentpay.discovery.v1 }
        sellerId: { type: string, pattern: '^sel_[A-Za-z0-9]+$' }
        sellerSlug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        availability: { type: string, const: inactive }
        reason: { type: string, enum: [suspended, cancelled, closed] }
        publicationRevision: { type: integer, minimum: 1 }
        issuedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        canonicalOrigin: { type: string, format: uri }
    DiscoverySignature:
      type: object
      additionalProperties: false
      required: [alg, kid, canonicalization, domainSeparator, value]
      properties:
        alg: { type: string, const: ES256 }
        kid: { type: string, minLength: 1, maxLength: 128 }
        canonicalization: { type: string, const: RFC8785 }
        domainSeparator: { type: string, const: agentpay.discovery.v1 }
        value: { type: string, minLength: 86, maxLength: 86, description: 'Unpadded base64url encoding of the raw 64-byte ES256 signature over agentpay.discovery.v1, one NUL byte, and RFC 8785 canonical document bytes.' }
    SignedStorefrontManifest:
      type: object
      additionalProperties: false
      required: [document, signature]
      properties:
        document: { $ref: '#/components/schemas/StorefrontManifest' }
        signature: { $ref: '#/components/schemas/DiscoverySignature' }
    SignedStorefrontTombstone:
      type: object
      additionalProperties: false
      required: [document, signature]
      properties:
        document: { $ref: '#/components/schemas/StorefrontTombstone' }
        signature: { $ref: '#/components/schemas/DiscoverySignature' }
    SignedPublicProductDocument:
      type: object
      additionalProperties: false
      required: [document, signature]
      properties:
        document: { $ref: '#/components/schemas/PublicProductDocument' }
        signature: { $ref: '#/components/schemas/DiscoverySignature' }
    ErrorCode:
      type: string
      enum:
        - invalid_request
        - invalid_credential
        - token_expired
        - token_revoked
        - subscription_inactive
        - insufficient_scope
        - permission_denied
        - not_found
        - seller_inactive
        - invitation_expired
        - state_conflict
        - idempotency_conflict
        - payment_required
        - payment_rejected
        - payment_replayed
        - token_replayed
        - approval_required
        - validation_failed
        - rate_limited
        - dependency_unavailable
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            requestId: { type: string }
            details: { type: object }
    ApprovalRequiredError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId, details]
          properties:
            code: { type: string, const: approval_required }
            message: { type: string }
            requestId: { type: string }
            details:
              type: object
              additionalProperties: false
              required: [approvalSessionId]
              properties:
                approvalSessionId: { type: string, pattern: '^aps_[A-Za-z0-9]+$' }
    AtomicAmount: { type: string, pattern: '^[0-9]+$', example: '10000' }
...
    Sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
  responses:
...
    ProxiedResponse:
      description: Successful seller response
      headers:
        X-AgentPay-Transaction-Id: { schema: { type: string } }
        PAYMENT-RESPONSE: { schema: { type: string }, description: x402 v2 settlement response when supplied by the payment adapter }
      content: { application/json: { schema: {} } }
    PaymentRequired:
      description: x402 v2 payment challenge
      headers:
        PAYMENT-REQUIRED: { required: true, schema: { type: string }, description: Base64-encoded x402 PaymentRequired payload }
        Cache-Control: { schema: { type: string, const: no-store } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ApprovalRequired:
      description: Purchase requires approval before a payment challenge can be issued
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApprovalRequiredError' }
    BadRequest: { description: 400 invalid_request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized:
      description: 401 invalid_credential, token_expired, or token_revoked
      headers:
        WWW-Authenticate: { schema: { type: string } }
        Cache-Control: { schema: { type: string, const: no-store } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound: { description: '404 not_found, including concealed cross-tenant resources', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: '409 state_conflict, idempotency_conflict, payment_replayed, or token_replayed', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    SellerInactive: { description: 410 seller_inactive; known seller cannot participate in discovery or new commerce, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unprocessable: { description: 422 validation_failed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    PermissionDenied: { description: '403 subscription_inactive, insufficient_scope, or permission_denied', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    RateLimited:
      description: Caller, credential, browser session, seller quota, or public endpoint rate limit is exhausted
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    DependencyUnavailable:
      description: 503 dependency_unavailable; authoritative authorization or transaction state cannot be established
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }